Refactored Environment class
