+++ systemtest/src/main/java/io/strimzi/systemtest/Resources.java
@@ -59,6 +59,7 @@ import io.strimzi.api.kafka.model.KafkaUserScramSha512ClientAuthentication;
 import io.strimzi.api.kafka.model.KafkaUserTlsClientAuthentication;
 import io.strimzi.api.kafka.model.SingleVolumeStorage;
 import io.strimzi.systemtest.utils.StUtils;
+import io.strimzi.test.Environment;
 import io.strimzi.test.TestUtils;
 import io.strimzi.test.k8s.KubeClient;
 import org.apache.kafka.clients.consumer.ConsumerConfig;
