+++ systemtest/src/test/java/io/strimzi/systemtest/KafkaST.java
@@ -25,6 +25,7 @@ import io.strimzi.api.kafka.model.listener.KafkaListenerPlain;
 import io.strimzi.api.kafka.model.listener.KafkaListenerTls;
 import io.strimzi.systemtest.annotations.OpenShiftOnly;
 import io.strimzi.systemtest.utils.StUtils;
+import io.strimzi.test.Environment;
 import io.strimzi.test.TestUtils;
 import io.strimzi.test.k8s.Oc;
 import io.strimzi.test.timemeasuring.Operation;
